backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
mysql-connector-python==8.0.29
python-dotenv==0.19.0
pandas==1.4.3
reportlab==3.6.11
pyjwt==2.4.0
bcrypt==3.2.0